Moved the tests to the bottom of the config_handler.rs file, removed config_handler_tests.rs file as per requested
