+++ saito-rust/src/test/mod.rs
@@ -8,4 +8,3 @@ mod test_io_handler;
 pub mod test_manager;
 mod test_setup;
 mod wallet_tests;
-mod config_handler_tests;
